+++ ReadQiitaApp_iOS/Screen/ArticleView.swift
@@ -9,7 +9,7 @@ import SwiftUI
 import ComposableArchitecture
 
 
-
+// MARK: - Reducer
 
 struct ArticleReducer: Reducer {
     struct State: Equatable {
